Get rid of charmap option on font test & add FPS display
I also got rid of the optimization where we only render a character
if it's different from the character previously rendered (in the back buffer)
It seemed to work most of the time, with the occasional quirky
behavior. The frame rate dropped from ~4000 FPS to ~1000 FPS as
a result. I can live with that.
I'm keeping the double-buffering for now even though it may seem
pointless without this optimization. I may revisit it later and
I have some ideas for a cool visual effect it may enable.